feat: 시간 차이를 나타내는 getTimeAgo 함수를 유틸 함수로 분리하여 처리 후 필요한 위치에 적용
